Finished all the commenting
