+++ common/src/main/java/software/bluelib/test/utils/MessageUtils.java
@@ -2,38 +2,51 @@
 
 package software.bluelib.test.utils;
 
-import java.util.List;
 import net.minecraft.gametest.framework.GameTestHelper;
 import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerPlayer;
 import software.bluelib.markdown.MarkdownParser;
 
+import java.util.List;
+
 /**
- * A {@code public class} that provides utility methods for tests related to messages and chat.
+ * Utility class for sending messages to players and generating random hex values.
+ * <p>
+ * Purpose: This class provides utility methods for sending formatted messages to players and generating random hex values.<br>
+ * When: Used during game tests to communicate with players and generate random values.<br>
+ * Where: Invoked in game test scenarios and other utility contexts.<br>
+ * Additional Info: The messages are formatted using the {@link MarkdownParser} class.
+ * </p>
  * Key Methods:
  * <ul>
- * <li>{@link #sendMessageToPlayers(GameTestHelper, String)} - Sends a formatted message to all players on the server.</li>
- * <li>{@link #getRandomHex()} - Generates a random six-digit hexadecimal value.</li>
+ * <li>{@link #sendMessageToPlayers(GameTestHelper, String)} - Sends a formatted message to all players.</li>
+ * <li>{@link #getRandomHex()} - Generates a random hex value.</li>
  * </ul>
  *
  * @author MeAlam
  * @version 1.6.0
+ * @see MarkdownParser
+ * @see GameTestHelper
+ * @see ServerPlayer
+ * @see Component
  * @since 1.6.0
  */
 public class MessageUtils {
 
     /**
-     * A {@code public static} method that sends a formatted message to all players on the server.
+     * Sends a formatted message to all players in the game test.
      * <p>
-     * This method retrieves the list of players from the {@link GameTestHelper} and applies Markdown
-     * formatting to the message using the {@link MarkdownParser#parseMarkdown(Component)} method. If
-     * no players are found or the message formatting fails, the test will fail with an appropriate
-     * error message.
+     * Purpose: This method sends a message to all players after formatting it using Markdown.<br>
+     * When: Called during game tests to communicate with players.<br>
+     * Where: Used in game test scenarios where player communication is required.<br>
+     * Additional Info: If no players are found or the message formatting fails, the test will fail.
      * </p>
      *
-     * @param pHelper  {@link GameTestHelper} - The game test helper providing context and utilities.
-     * @param pMessage {@link String} - The message to be sent to players.
+     * @param pHelper  The game test helper instance.
+     * @param pMessage The message to be sent to players.
      * @author MeAlam
+     * @see MarkdownParser
+     * @see GameTestHelper
      * @since 1.6.0
      */
     public static void sendMessageToPlayers(GameTestHelper pHelper, String pMessage) {
@@ -54,14 +67,17 @@ public class MessageUtils {
     }
 
     /**
-     * A {@code public static} method that generates a random six-digit hexadecimal value.
+     * Generates a random hex value.
      * <p>
-     * This method creates a random integer between 100,000 and 999,999, simulating the range of a
-     * six-digit hexadecimal number.
+     * Purpose: This method generates a random hex value within a specified range.<br>
+     * When: Called when a random hex value is needed.<br>
+     * Where: Used in various utility contexts where random values are required.<br>
+     * Additional Info: The generated hex value is between 100000 and 999999.
      * </p>
      *
-     * @return {@link Integer} - A randomly generated six-digit hexadecimal value.
+     * @return A random hex value.
      * @author MeAlam
+     * @see java.util.Random
      * @since 1.6.0
      */
     public static int getRandomHex() {
